Move some classes to Lib.Config namespace
- Move ExtensionMethods
- Move ValidatingNodeDeserializer
- Prettified ExtensionMethods
